Add describer for the Structured Worker.
