+++ packages/agent-kit/src/index.ts
@@ -8,10 +8,13 @@ import { GraphToKitAdapter, KitBuilder } from "@google-labs/breadboard/kits";
 
 import kit from "./kit.js";
 import { addKit } from "@google-labs/breadboard";
-import { WorkerType } from "./boards/worker.js";
-import { addDescriber, workerDescriber } from "./hacks.js";
+import { WorkerType, workerDescriber } from "./boards/worker.js";
+import { addDescriber } from "./hacks.js";
 import { RepeaterType } from "./boards/repeater.js";
-import { StructuredWorkerType } from "./boards/structured-worker.js";
+import {
+  StructuredWorkerType,
+  structuredWorkerDescriber,
+} from "./boards/structured-worker.js";
 
 // TODO: Replace with the actual URL.
 const KIT_BASE_URL =
@@ -32,7 +35,10 @@ const builder = new KitBuilder(
 const AgentKit = builder.build({
   worker: addDescriber(adapter.handlerForNode("worker"), workerDescriber),
   repeater: adapter.handlerForNode("repeater"),
-  structuredWorker: adapter.handlerForNode("structured-worker"),
+  structuredWorker: addDescriber(
+    adapter.handlerForNode("structured-worker"),
+    structuredWorkerDescriber
+  ),
 });
 
 export type AgentKit = InstanceType<typeof AgentKit>;
