+++ packages/breadboard-web/public/graphs/ad-writer.json
@@ -259,12 +259,13 @@
           "type": "object",
           "properties": {
             "context": {
+              "type": "string",
               "title": "Ad specs",
+              "description": "The context to use for the worker",
               "format": "multiline",
               "examples": [
                 "This ad is for my lawn care company that will fit into an inch of newspaper copy. It's called \"Max's Lawn Care\" and it should use the slogan \"I care about your lawn.\" Emphasize the folksiness of it being a local, sole proprietorship that I started after graduating from high school."
-              ],
-              "type": "string"
+              ]
             }
           },
           "required": [
