+++ packages/breadboard-web/public/graphs/json-agent.json
@@ -44,11 +44,13 @@
           "properties": {
             "context": {
               "type": "string",
-              "title": "context"
+              "title": "Context",
+              "description": "The context after generation"
             },
             "json": {
               "type": "string",
-              "title": "json"
+              "title": "Output",
+              "description": "The output from the agent. Use this to just get the output without any previous context."
             }
           },
           "required": [
@@ -71,27 +73,30 @@
           "type": "object",
           "properties": {
             "context": {
-              "title": "Context",
               "type": "array",
+              "title": "Context",
+              "description": "The context to use for the worker",
               "examples": [
                 "[]"
               ]
             },
             "instruction": {
+              "type": "string",
               "title": "Instruction",
+              "description": "The instruction we want to give to the worker so that shapes its character and orients it a bit toward the task we want to give it.",
               "examples": [
                 "You are building a team of skilled experts to create high quality rhyming poems that will be used as lyrics for jingles in TV commercials. These experts can only read text and produce text. Creating melodies and producing music is not their responsibility. The experts will work as a team, collaborating, creating, reviewing, critiquing, and iteratively improving the quality of the poems.\n\nPlease identify the necessary job descriptions of these experts."
               ],
-              "format": "multiline",
-              "type": "string"
+              "format": "multiline"
             },
             "schema": {
+              "type": "object",
               "title": "Schema",
+              "format": "multiline",
+              "description": "The JSON schema to use for the worker",
               "examples": [
                 "{\n  \"type\": \"object\",\n  \"properties\": {\n    \"descriptions\": {\n      \"type\": \"array\",\n      \"items\": {\n        \"type\": \"object\",\n        \"properties\": {\n          \"title\": {\n            \"type\": \"string\",\n            \"description\": \"expert's title\"\n          },\n          \"responsibilities\": {\n            \"type\": \"array\",\n            \"items\": {\n              \"type\": \"string\",\n              \"description\": \"expert's responsibilities\"\n            }\n          }\n        }\n      }\n    }\n  }\n}"
               ],
-              "type": "object",
-              "format": "multiline",
               "default": "{}"
             }
           },
